+++ sync-for-payables/src/funcs/billPaymentsCreate.ts
@@ -3,12 +3,9 @@
  */
 
 import { CodatSyncPayablesCore } from "../core.js";
-import {
-  encodeJSON as encodeJSON$,
-  encodeSimple as encodeSimple$,
-} from "../lib/encodings.js";
-import * as m$ from "../lib/matchers.js";
-import * as schemas$ from "../lib/schemas.js";
+import { encodeJSON, encodeSimple } from "../lib/encodings.js";
+import * as M from "../lib/matchers.js";
+import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
@@ -35,7 +32,7 @@ import { Result } from "../sdk/types/fp.js";
  * [Bill payments](https://docs.codat.io/sync-for-payables-api#/schemas/BillPayment) are an allocation of money within any Accounts Payable account.
  */
 export async function billPaymentsCreate(
-  client$: CodatSyncPayablesCore,
+  client: CodatSyncPayablesCore,
   request: operations.CreateBillPaymentRequest,
   options?: RequestOptions,
 ): Promise<
@@ -51,74 +48,71 @@ export async function billPaymentsCreate(
     | ConnectionError
   >
 > {
-  const input$ = request;
-
-  const parsed$ = schemas$.safeParse(
-    input$,
-    (value$) =>
-      operations.CreateBillPaymentRequest$outboundSchema.parse(value$),
+  const parsed = safeParse(
+    request,
+    (value) => operations.CreateBillPaymentRequest$outboundSchema.parse(value),
     "Input validation failed",
   );
-  if (!parsed$.ok) {
-    return parsed$;
+  if (!parsed.ok) {
+    return parsed;
   }
-  const payload$ = parsed$.value;
-  const body$ = encodeJSON$("body", payload$.billPaymentPrototype, {
+  const payload = parsed.value;
+  const body = encodeJSON("body", payload.billPaymentPrototype, {
     explode: true,
   });
 
-  const pathParams$ = {
-    billId: encodeSimple$("billId", payload$.billId, {
+  const pathParams = {
+    billId: encodeSimple("billId", payload.billId, {
       explode: false,
       charEncoding: "percent",
     }),
-    companyId: encodeSimple$("companyId", payload$.companyId, {
+    companyId: encodeSimple("companyId", payload.companyId, {
       explode: false,
       charEncoding: "percent",
     }),
-    connectionId: encodeSimple$("connectionId", payload$.connectionId, {
+    connectionId: encodeSimple("connectionId", payload.connectionId, {
       explode: false,
       charEncoding: "percent",
     }),
   };
 
-  const path$ = pathToFunc(
+  const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/payables/bills/{billId}/payment",
-  )(pathParams$);
+  )(pathParams);
 
-  const headers$ = new Headers({
+  const headers = new Headers({
     "Content-Type": "application/json",
     Accept: "application/json",
-    "Idempotency-Key": encodeSimple$(
+    "Idempotency-Key": encodeSimple(
       "Idempotency-Key",
-      payload$["Idempotency-Key"],
+      payload["Idempotency-Key"],
       { explode: false, charEncoding: "none" },
     ),
   });
 
-  const authHeader$ = await extractSecurity(client$.options$.authHeader);
-  const security$ = authHeader$ == null ? {} : { authHeader: authHeader$ };
+  const secConfig = await extractSecurity(client._options.authHeader);
+  const securityInput = secConfig == null ? {} : { authHeader: secConfig };
   const context = {
     operationID: "create-bill-payment",
     oAuth2Scopes: [],
-    securitySource: client$.options$.authHeader,
+    securitySource: client._options.authHeader,
   };
-  const securitySettings$ = resolveGlobalSecurity(security$);
+  const requestSecurity = resolveGlobalSecurity(securityInput);
 
-  const requestRes = client$.createRequest$(context, {
-    security: securitySettings$,
+  const requestRes = client._createRequest(context, {
+    security: requestSecurity,
     method: "POST",
-    path: path$,
-    headers: headers$,
-    body: body$,
-    timeoutMs: options?.timeoutMs || client$.options$.timeoutMs || -1,
+    path: path,
+    headers: headers,
+    body: body,
+    timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);
   if (!requestRes.ok) {
     return requestRes;
   }
-  const request$ = requestRes.value;
+  const req = requestRes.value;
 
-  const doResult = await client$.do$(request$, {
+  const doResult = await client._do(req, {
     context,
     errorCodes: [
       "400",
@@ -134,7 +128,7 @@ export async function billPaymentsCreate(
       "5XX",
     ],
     retryConfig: options?.retries
-      || client$.options$.retryConfig
+      || client._options.retryConfig
       || {
         strategy: "backoff",
         backoff: {
@@ -152,11 +146,11 @@ export async function billPaymentsCreate(
   }
   const response = doResult.value;
 
-  const responseFields$ = {
-    HttpMeta: { Response: response, Request: request$ },
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
   };
 
-  const [result$] = await m$.match<
+  const [result] = await M.match<
     shared.BillPayment,
     | errors.ErrorMessage
     | SDKError
@@ -167,16 +161,16 @@ export async function billPaymentsCreate(
     | RequestTimeoutError
     | ConnectionError
   >(
-    m$.json(200, shared.BillPayment$inboundSchema),
-    m$.jsonErr(
+    M.json(200, shared.BillPayment$inboundSchema),
+    M.jsonErr(
       [400, 401, 402, 403, 404, 409, 429, 500, 503],
       errors.ErrorMessage$inboundSchema,
     ),
-    m$.fail(["4XX", "5XX"]),
-  )(response, { extraFields: responseFields$ });
-  if (!result$.ok) {
-    return result$;
+    M.fail(["4XX", "5XX"]),
+  )(response, { extraFields: responseFields });
+  if (!result.ok) {
+    return result;
   }
 
-  return result$;
+  return result;
 }