+++ sync-for-payables/src/funcs/billsListAttachments.ts
@@ -3,9 +3,9 @@
  */
 
 import { CodatSyncPayablesCore } from "../core.js";
-import { encodeSimple as encodeSimple$ } from "../lib/encodings.js";
-import * as m$ from "../lib/matchers.js";
-import * as schemas$ from "../lib/schemas.js";
+import { encodeSimple } from "../lib/encodings.js";
+import * as M from "../lib/matchers.js";
+import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
@@ -32,7 +32,7 @@ import { Result } from "../sdk/types/fp.js";
  * [Bills](https://docs.codat.io/sync-for-payables-api#/schemas/Bill) are invoices that represent the SMB's financial obligations to their supplier for a purchase of goods or services.
  */
 export async function billsListAttachments(
-  client$: CodatSyncPayablesCore,
+  client: CodatSyncPayablesCore,
   request: operations.ListBillAttachmentsRequest,
   options?: RequestOptions,
 ): Promise<
@@ -48,66 +48,64 @@ export async function billsListAttachments(
     | ConnectionError
   >
 > {
-  const input$ = request;
-
-  const parsed$ = schemas$.safeParse(
-    input$,
-    (value$) =>
-      operations.ListBillAttachmentsRequest$outboundSchema.parse(value$),
+  const parsed = safeParse(
+    request,
+    (value) =>
+      operations.ListBillAttachmentsRequest$outboundSchema.parse(value),
     "Input validation failed",
   );
-  if (!parsed$.ok) {
-    return parsed$;
+  if (!parsed.ok) {
+    return parsed;
   }
-  const payload$ = parsed$.value;
-  const body$ = null;
+  const payload = parsed.value;
+  const body = null;
 
-  const pathParams$ = {
-    billId: encodeSimple$("billId", payload$.billId, {
+  const pathParams = {
+    billId: encodeSimple("billId", payload.billId, {
       explode: false,
       charEncoding: "percent",
     }),
-    companyId: encodeSimple$("companyId", payload$.companyId, {
+    companyId: encodeSimple("companyId", payload.companyId, {
       explode: false,
       charEncoding: "percent",
     }),
-    connectionId: encodeSimple$("connectionId", payload$.connectionId, {
+    connectionId: encodeSimple("connectionId", payload.connectionId, {
       explode: false,
       charEncoding: "percent",
     }),
   };
 
-  const path$ = pathToFunc(
+  const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/bills/{billId}/attachments",
-  )(pathParams$);
+  )(pathParams);
 
-  const headers$ = new Headers({
+  const headers = new Headers({
     Accept: "application/json",
   });
 
-  const authHeader$ = await extractSecurity(client$.options$.authHeader);
-  const security$ = authHeader$ == null ? {} : { authHeader: authHeader$ };
+  const secConfig = await extractSecurity(client._options.authHeader);
+  const securityInput = secConfig == null ? {} : { authHeader: secConfig };
   const context = {
     operationID: "list-bill-attachments",
     oAuth2Scopes: [],
-    securitySource: client$.options$.authHeader,
+    securitySource: client._options.authHeader,
   };
-  const securitySettings$ = resolveGlobalSecurity(security$);
+  const requestSecurity = resolveGlobalSecurity(securityInput);
 
-  const requestRes = client$.createRequest$(context, {
-    security: securitySettings$,
+  const requestRes = client._createRequest(context, {
+    security: requestSecurity,
     method: "GET",
-    path: path$,
-    headers: headers$,
-    body: body$,
-    timeoutMs: options?.timeoutMs || client$.options$.timeoutMs || -1,
+    path: path,
+    headers: headers,
+    body: body,
+    timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);
   if (!requestRes.ok) {
     return requestRes;
   }
-  const request$ = requestRes.value;
+  const req = requestRes.value;
 
-  const doResult = await client$.do$(request$, {
+  const doResult = await client._do(req, {
     context,
     errorCodes: [
       "401",
@@ -122,7 +120,7 @@ export async function billsListAttachments(
       "5XX",
     ],
     retryConfig: options?.retries
-      || client$.options$.retryConfig
+      || client._options.retryConfig
       || {
         strategy: "backoff",
         backoff: {
@@ -140,11 +138,11 @@ export async function billsListAttachments(
   }
   const response = doResult.value;
 
-  const responseFields$ = {
-    HttpMeta: { Response: response, Request: request$ },
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
   };
 
-  const [result$] = await m$.match<
+  const [result] = await M.match<
     shared.Attachment,
     | errors.ErrorMessage
     | SDKError
@@ -155,16 +153,16 @@ export async function billsListAttachments(
     | RequestTimeoutError
     | ConnectionError
   >(
-    m$.json(200, shared.Attachment$inboundSchema),
-    m$.jsonErr(
+    M.json(200, shared.Attachment$inboundSchema),
+    M.jsonErr(
       [401, 402, 403, 404, 409, 429, 500, 503],
       errors.ErrorMessage$inboundSchema,
     ),
-    m$.fail(["4XX", "5XX"]),
-  )(response, { extraFields: responseFields$ });
-  if (!result$.ok) {
-    return result$;
+    M.fail(["4XX", "5XX"]),
+  )(response, { extraFields: responseFields });
+  if (!result.ok) {
+    return result;
   }
 
-  return result$;
+  return result;
 }