+++ sync-for-payables/src/funcs/companiesList.ts
@@ -3,9 +3,9 @@
  */
 
 import { CodatSyncPayablesCore } from "../core.js";
-import { encodeFormQuery as encodeFormQuery$ } from "../lib/encodings.js";
-import * as m$ from "../lib/matchers.js";
-import * as schemas$ from "../lib/schemas.js";
+import { encodeFormQuery } from "../lib/encodings.js";
+import * as M from "../lib/matchers.js";
+import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
@@ -33,8 +33,8 @@ import { Result } from "../sdk/types/fp.js";
  * Each company can have multiple [connections](https://docs.codat.io/sync-for-payables-api#/schemas/Connection) to different data sources, such as one connection to Xero for accounting data, two connections to Plaid for two bank accounts, and a connection to Zettle for POS data.
  */
 export async function companiesList(
-  client$: CodatSyncPayablesCore,
-  request: operations.ListCompaniesRequest,
+  client: CodatSyncPayablesCore,
+  request?: operations.ListCompaniesRequest | undefined,
   options?: RequestOptions,
 ): Promise<
   Result<
@@ -49,56 +49,55 @@ export async function companiesList(
     | ConnectionError
   >
 > {
-  const input$ = request;
-
-  const parsed$ = schemas$.safeParse(
-    input$,
-    (value$) => operations.ListCompaniesRequest$outboundSchema.parse(value$),
+  const parsed = safeParse(
+    request,
+    (value) =>
+      operations.ListCompaniesRequest$outboundSchema.optional().parse(value),
     "Input validation failed",
   );
-  if (!parsed$.ok) {
-    return parsed$;
+  if (!parsed.ok) {
+    return parsed;
   }
-  const payload$ = parsed$.value;
-  const body$ = null;
+  const payload = parsed.value;
+  const body = null;
 
-  const path$ = pathToFunc("/companies")();
+  const path = pathToFunc("/companies")();
 
-  const query$ = encodeFormQuery$({
-    "orderBy": payload$.orderBy,
-    "page": payload$.page,
-    "pageSize": payload$.pageSize,
-    "query": payload$.query,
+  const query = encodeFormQuery({
+    "orderBy": payload?.orderBy,
+    "page": payload?.page,
+    "pageSize": payload?.pageSize,
+    "query": payload?.query,
   });
 
-  const headers$ = new Headers({
+  const headers = new Headers({
     Accept: "application/json",
   });
 
-  const authHeader$ = await extractSecurity(client$.options$.authHeader);
-  const security$ = authHeader$ == null ? {} : { authHeader: authHeader$ };
+  const secConfig = await extractSecurity(client._options.authHeader);
+  const securityInput = secConfig == null ? {} : { authHeader: secConfig };
   const context = {
     operationID: "list-companies",
     oAuth2Scopes: [],
-    securitySource: client$.options$.authHeader,
+    securitySource: client._options.authHeader,
   };
-  const securitySettings$ = resolveGlobalSecurity(security$);
+  const requestSecurity = resolveGlobalSecurity(securityInput);
 
-  const requestRes = client$.createRequest$(context, {
-    security: securitySettings$,
+  const requestRes = client._createRequest(context, {
+    security: requestSecurity,
     method: "GET",
-    path: path$,
-    headers: headers$,
-    query: query$,
-    body: body$,
-    timeoutMs: options?.timeoutMs || client$.options$.timeoutMs || -1,
+    path: path,
+    headers: headers,
+    query: query,
+    body: body,
+    timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);
   if (!requestRes.ok) {
     return requestRes;
   }
-  const request$ = requestRes.value;
+  const req = requestRes.value;
 
-  const doResult = await client$.do$(request$, {
+  const doResult = await client._do(req, {
     context,
     errorCodes: [
       "400",
@@ -113,7 +112,7 @@ export async function companiesList(
       "5XX",
     ],
     retryConfig: options?.retries
-      || client$.options$.retryConfig
+      || client._options.retryConfig
       || {
         strategy: "backoff",
         backoff: {
@@ -131,11 +130,11 @@ export async function companiesList(
   }
   const response = doResult.value;
 
-  const responseFields$ = {
-    HttpMeta: { Response: response, Request: request$ },
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
   };
 
-  const [result$] = await m$.match<
+  const [result] = await M.match<
     shared.Companies,
     | errors.ErrorMessage
     | SDKError
@@ -146,16 +145,16 @@ export async function companiesList(
     | RequestTimeoutError
     | ConnectionError
   >(
-    m$.json(200, shared.Companies$inboundSchema),
-    m$.jsonErr(
+    M.json(200, shared.Companies$inboundSchema),
+    M.jsonErr(
       [400, 401, 402, 403, 404, 429, 500, 503],
       errors.ErrorMessage$inboundSchema,
     ),
-    m$.fail(["4XX", "5XX"]),
-  )(response, { extraFields: responseFields$ });
-  if (!result$.ok) {
-    return result$;
+    M.fail(["4XX", "5XX"]),
+  )(response, { extraFields: responseFields });
+  if (!result.ok) {
+    return result;
   }
 
-  return result$;
+  return result;
 }