+++ sync-for-payables/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v4.0.0] sync-for-payables
 ### Releases
 - [NPM v4.0.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/4.0.0 - sync-for-payables
+
+## 2024-10-02 13:11:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.405.9 (2.429.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.0.0] sync-for-payables
+### Releases
+- [NPM v5.0.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/5.0.0 - sync-for-payables